fixing importations

+++ src/erase/erase.service.ts
@@ -1,6 +1,5 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { PrismaService } from 'src/prisma/prisma.service';
-import { UsersRepositories } from 'src/users/users.repositories';
+import { UsersRepositories } from '../users/users.repositories';
 import * as bcrypt from "bcrypt"
 
 @Injectable()
@@ -10,7 +9,7 @@ export class EraseService {
     async deleteAccount(password: string,userId: number){
         const user = await this.usersRepositories.findById(userId);
         const isValid = await bcrypt.compare(password,user.password)
-        //if(!isValid) throw new UnauthorizedException("invalid password");
+        if(!isValid) throw new UnauthorizedException("invalid password");
         await this.usersRepositories.deleteAll(userId);
     }
 }
